rpm: retry installing the package if the mirror server is busy
When installing some packages, if the mirror server failed with
503 code, which means the mirror server temporarily not available,
we should retry it later. But the yum tool just skips it and
retries other mirrors, which may not contain them.

For the cephfs suites, there maybe will fire the many test cases
at the same time, and for each test case it may fire several nodes
to install tens of packages at the same time. This may cause the
mirror server overloaded.

We need one safe method to retry it.

Fixes: https://tracker.ceph.com/issues/47166
Signed-off-by: Xiubo Li <xiubli@redhat.com>